requirements.txt
browser-use
python-dotenv
